fix: todo 길이제한 100
- TodoItem 분리
